+++ src/Settings/hooks/index.js
@@ -1,0 +1,1 @@
+export { useBankingInformation } from './useBankingInformation';