+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.1.0 (IN PROGRESS)
 
 * Designate Organization as donor. Refs UIORGS-383.
+* Settings for banking information. Refs UIORGS-391.
 
 ## [5.0.0](https://github.com/folio-org/ui-organizations/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v4.0.0...v5.0.0)